Change the type of the counter variable in for loops

+++ SelectionSort.tpp
@@ -1,10 +1,10 @@
 template <class TP>
 void sortPersons(const int pCount, TP** persons) {
     TP* tmp;
-    int minNumber = 0;
-    for (int tNumber = 0; tNumber < pCount; tNumber++) {
+    size_t minNumber = 0;
+    for (size_t tNumber = 0; tNumber < pCount; tNumber++) {
         minNumber = tNumber;
-        for (int nNumber = tNumber + 1; nNumber < pCount; nNumber++) {
+        for (size_t nNumber = tNumber + 1; nNumber < pCount; nNumber++) {
             minNumber = (persons[nNumber]->isLess(persons[minNumber])) ? nNumber : minNumber;
         }
         if (tNumber != minNumber) {
